Add pdfminer for PDF text extraction
