+++ watermark_remover/tools/scrape.py
@@ -1,32 +1,11 @@
-"""Tool definitions for the Watermark Remover agents.
+"""Scraping tool and shared utilities for the Watermark Remover project.
 
-Each function in this module is decorated with the ``langchain.agents.tool``
-decorator so that it can be invoked by LangChain agents.  The tools provide
-an abstraction over the core functionalities of the Watermark Remover
-project:
-
-* ``scrape_music``: returns a directory containing sheet music images.  In
-  this proof‑of‑concept implementation, no real scraping occurs; the
-  function merely returns a pre‑existing directory.  It is left as a stub
-  to be replaced with Selenium or API logic as needed.
-
-* ``remove_watermark``: loads a U‑Net model and applies it to each image
-  found in the input directory, saving the watermark‑free versions to a new
-  directory.
-
-* ``upscale_images``: loads a VDSR model and applies it to each image
-  found in the input directory, saving the upscaled results to a new
-  directory.
-
-* ``assemble_pdf``: collects all images from a directory and assembles
-  them into a multi‑page PDF.
-
-All tools return the path to the directory or file they produce.  If the
-specified model directory does not contain weights, the ``load_best_model``
-function silently fails and the model runs with randomly initialised weights;
-this keeps the example self‑contained and avoids bundling large model
-weights in the repository.  Users can populate the ``models/`` directories
-with their own trained checkpoints to achieve meaningful results.
+This module exposes the :func:`scrape_music` LangChain tool which locates
+sheet‑music images on disk or, when necessary, downloads them from the
+internet.  It also houses helper functions and module‑level state used by
+the other tools, such as :func:`sanitize_title`, ``SCRAPE_METADATA`` and the
+root ``logger``.  Subsequent modules import these objects to ensure
+consistent logging and directory naming across the pipeline.
 """
 
 from __future__ import annotations
@@ -642,27 +621,6 @@ def sanitize_title(value: str) -> str:
     sanitized = re.sub(r"_+", "_", sanitized)
     return sanitized.strip("_")
 
-# Import model definitions lazily.  These imports can be heavy and
-# require optional dependencies such as torch, torchvision and
-# pytorch_msssim.  To avoid import errors when those libraries are not
-# installed, we catch ImportError and provide a helpful message at
-# runtime.
-try:
-    import torch  # type: ignore
-    from watermark_remover.inference.model_functions import (
-        UNet,
-        VDSR,
-        PIL_to_tensor,
-        tensor_to_PIL,
-        load_best_model,
-    )
-except Exception as e:  # broad except to handle ImportError and others
-    UNet = VDSR = None  # type: ignore
-    PIL_to_tensor = tensor_to_PIL = load_best_model = None  # type: ignore
-    torch = None  # type: ignore
-    _import_error = e
-else:
-    _import_error = None
 
 # Set up a module‑level logger.  The log level can be controlled via the
 # LOG_LEVEL environment variable (e.g. export LOG_LEVEL=DEBUG).  If not set,
@@ -1939,349 +1897,3 @@ def _scrape_with_selenium(title: str, instrument: str, key: str, artist: Optiona
             pass
 
 
-@tool
-def remove_watermark(input_dir: str, model_dir: str = "models/Watermark_Removal", output_dir: str = "processed") -> str:
-    """Remove watermarks from all images in ``input_dir`` using a UNet model.
-
-    Parameters
-    ----------
-    input_dir : str
-        Directory containing input images.
-    model_dir : str
-        Directory containing UNet checkpoint files.  The most recently
-        trained model will be selected based on filename ordering.
-    output_dir : str
-        Directory to which watermark‑free images are saved.
-
-    Returns
-    -------
-    str
-        Path to the directory containing watermark‑free images.
-
-    Notes
-    -----
-    If PyTorch or the UNet implementation cannot be imported, this function
-    will raise an ImportError when called.  If the model directory
-    contains no checkpoints, the UNet model will run with randomly
-    initialised weights, which will not produce meaningful results but
-    allows the pipeline to run end‑to‑end without bundling large model
-    weights in the repository.
-    """
-    start = time.perf_counter()
-    if _import_error is not None:
-        raise ImportError(
-            f"Cannot import UNet and related utilities: {_import_error}. "
-            "Ensure torch and watermark_remover dependencies are installed."
-        )
-    if not os.path.isdir(input_dir):
-        raise FileNotFoundError(f"Input directory {input_dir} does not exist.")
-    if not os.path.isdir(model_dir):
-        raise FileNotFoundError(f"Model directory {model_dir} does not exist.")
-    # Determine the output directory.  If the caller provides a value
-    # other than the default, honour it.  Otherwise, derive a
-    # ``2_watermark_removed`` sibling inside the same run directory as
-    # the input.  This ensures that all intermediate artefacts live
-    # alongside the original images for this run.
-    if output_dir == "processed" or not output_dir:
-        parent_dir = os.path.dirname(input_dir.rstrip(os.sep))
-        output_dir = os.path.join(parent_dir, "2_watermark_removed")
-    os.makedirs(output_dir, exist_ok=True)
-    # List images to process
-    images = [f for f in os.listdir(input_dir) if f.lower().endswith((".png", ".jpg", ".jpeg", ".tif", ".tiff"))]
-    if not images:
-        raise RuntimeError(f"WMR: no images found in {input_dir}")
-    device = torch.device("cuda" if torch and torch.cuda.is_available() else "cpu")
-    model = UNet().to(device)
-    # Load the best checkpoint if available.  If loading fails, the model
-    # will continue with random weights so the pipeline still runs.
-    try:
-        load_best_model(model, model_dir)  # type: ignore[misc]
-    except Exception:
-        logger.warning("WMR: failed to load checkpoints from %s; using random weights", model_dir)
-    model.eval()
-    processed_dir = output_dir
-    for fname in images:
-        inp_path = os.path.join(input_dir, fname)
-        out_path = os.path.join(processed_dir, fname)
-        with torch.no_grad():
-            # Load and convert the image to a tensor.  If the tensor has three
-            # channels, explicitly convert the image to grayscale first to
-            # satisfy the UNet input shape (1 channel).  Some downloaded
-            # previews are RGB even when PIL_to_tensor normally converts
-            # images to grayscale.  Converting here ensures the network
-            # receives a 1×H×W tensor and avoids "expected 1 channel but got 3"
-            # runtime errors.
-            tensor = PIL_to_tensor(inp_path)
-            # tensor shape is (C, H, W)
-            if tensor.dim() == 3 and tensor.shape[0] != 1:
-                try:
-                    im = Image.open(inp_path).convert("L")
-                    # save temporary grayscale image
-                    tmp_path = inp_path + ".gray.tmp"
-                    im.save(tmp_path)
-                    tensor = PIL_to_tensor(tmp_path)
-                    os.remove(tmp_path)
-                except Exception:
-                    # fallback: average channels
-                    tensor = tensor.mean(dim=0, keepdim=True)
-            tensor = tensor.unsqueeze(0).to(device)
-            output = model(tensor)
-            img = tensor_to_PIL(output.squeeze(0).cpu())
-        os.makedirs(processed_dir, exist_ok=True)
-        img.save(out_path)
-        logger.info("WMR: processed %s -> %s", inp_path, out_path)
-    logger.info("WMR completed in %.3fs", time.perf_counter() - start)
-    # Do not track the processed directory for cleanup.  We intentionally
-    # preserve each intermediate stage under its timestamped run
-    # directory for debugging and reproducibility.
-    return processed_dir
-
-
-@tool
-def upscale_images(input_dir: str, model_dir: str = "models/VDSR", output_dir: str = "upscaled") -> str:
-    """Upscale all images in ``input_dir`` using a VDSR model.
-
-    This implementation mirrors the patch‑based algorithm used in the
-    original Watermark Remover project.  Each image is first upsampled
-    to a fixed size using nearest‑neighbour interpolation and then
-    processed in overlapping patches through the VDSR network.  The
-    results are stitched back together to form the final high‑resolution
-    image.
-
-    Parameters
-    ----------
-    input_dir : str
-        Directory containing watermark‑free images.
-    model_dir : str
-        Directory containing VDSR checkpoint files.
-    output_dir : str
-        Directory to which upscaled images are saved.
-
-    Returns
-    -------
-    str
-        Path to the directory containing upscaled images.
-
-    Notes
-    -----
-    If PyTorch or the VDSR implementation cannot be imported, this function
-    will raise an ImportError when called.  If the model directory
-    contains no checkpoints, the VDSR model will run with randomly
-    initialised weights, which will not produce meaningful results but
-    allows the pipeline to run end‑to‑end without bundling large model
-    weights in the repository.
-    """
-    start = time.perf_counter()
-    if _import_error is not None:
-        raise ImportError(
-            f"Cannot import VDSR and related utilities: {_import_error}. "
-            "Ensure torch and watermark_remover dependencies are installed."
-        )
-    if not os.path.isdir(input_dir):
-        raise FileNotFoundError(f"Input directory {input_dir} does not exist.")
-    if not os.path.isdir(model_dir):
-        raise FileNotFoundError(f"Model directory {model_dir} does not exist.")
-    # Determine the output directory.  If the caller provides a value
-    # other than the default, honour it.  Otherwise, derive a
-    # ``3_upscaled`` sibling inside the same run directory as the input.
-    if output_dir == "upscaled" or not output_dir:
-        parent_dir = os.path.dirname(input_dir.rstrip(os.sep))
-        output_dir = os.path.join(parent_dir, "3_upscaled")
-    os.makedirs(output_dir, exist_ok=True)
-    images = [
-        f
-        for f in os.listdir(input_dir)
-        if f.lower().endswith((".png", ".jpg", ".jpeg", ".tif", ".tiff"))
-    ]
-    if not images:
-        raise RuntimeError(f"UPSCALE: no images found in {input_dir}")
-    device = torch.device("cuda" if torch and torch.cuda.is_available() else "cpu")
-    # Instantiate the VDSR model and load the best checkpoint if available.
-    us_model = VDSR().to(device)
-    try:
-        load_best_model(us_model, model_dir)  # type: ignore[misc]
-    except Exception:
-        logger.warning("VDSR: failed to load checkpoints from %s; using random weights", model_dir)
-    us_model.eval()
-    # Define upsample operation to enlarge each image to the canonical size.
-    image_base_width, image_base_height = 1700, 2200
-    upsample = torch.nn.Upsample(size=(image_base_height, image_base_width), mode="nearest")
-    # Patch parameters matching the original implementation
-    padding_size = 16
-    patch_height = 550
-    patch_width = 850
-    # Process each image
-    for fname in images:
-        inp_path = os.path.join(input_dir, fname)
-        out_path = os.path.join(output_dir, fname)
-        # Convert image to tensor and move to device
-        with torch.no_grad():
-            tensor = PIL_to_tensor(inp_path)
-            tensor = tensor.unsqueeze(0).to(device)
-            # Upsample to target size
-            wm_output_upscaled = upsample(tensor)
-            # Pad so patches at the edges can be processed uniformly
-            padding = (padding_size, padding_size, padding_size, padding_size)
-            wm_output_upscaled_padded = torch.nn.functional.pad(
-                wm_output_upscaled, padding, value=1.0
-            )
-            # Prepare an output tensor
-            us_output = torch.zeros_like(wm_output_upscaled)
-            # Slide window over the upscaled image
-            for i in range(0, wm_output_upscaled.shape[-2], patch_height):
-                for j in range(0, wm_output_upscaled.shape[-1], patch_width):
-                    patch = wm_output_upscaled_padded[
-                        :,
-                        :,
-                        i : i + patch_height + padding_size * 2,
-                        j : j + patch_width + padding_size * 2,
-                    ]
-                    # Run VDSR on the patch
-                    us_patch = us_model(patch)
-                    # Remove padding from the output patch
-                    us_patch = us_patch[
-                        :,
-                        :,
-                        padding_size : -padding_size,
-                        padding_size : -padding_size,
-                    ]
-                    # Place the processed patch back into the output tensor
-                    us_output[
-                        :,
-                        :,
-                        i : i + patch_height,
-                        j : j + patch_width,
-                    ] = us_patch
-            # Convert the output tensor back to a PIL image and save
-            img = tensor_to_PIL(us_output.squeeze(0).cpu())
-            os.makedirs(output_dir, exist_ok=True)
-            img.save(out_path)
-            logger.info("UPSCALE: processed %s -> %s", inp_path, out_path)
-    logger.info("UPSCALE completed in %.3fs", time.perf_counter() - start)
-    # Do not track the upscaled directory for cleanup.  We preserve
-    # intermediate stages for debugging and reproducibility.
-    return output_dir
-
-
-@tool
-def assemble_pdf(image_dir: str, output_pdf: str = "output/output.pdf") -> str:
-    """Assemble images from a directory into a single PDF file.
-
-    Parameters
-    ----------
-    image_dir : str
-        Directory containing images to assemble into a PDF.
-    output_pdf : str
-        Name or path of the output PDF file.  If not provided, defaults to
-        ``"output/output.pdf"`` so that PDFs are written to the ``output``
-        directory relative to the current working directory.  If a
-        directory component is included in ``output_pdf``, it will be
-        created automatically.
-
-    Returns
-    -------
-    str
-        Path to the created PDF file.
-
-    Notes
-    -----
-    If reportlab cannot be imported, this function will raise an
-    ImportError.  The PDF will contain one page per image, preserving
-    the original aspect ratio.
-    """
-    start = time.perf_counter()
-    if not os.path.isdir(image_dir):
-        raise FileNotFoundError(f"Image directory {image_dir} does not exist.")
-    try:
-        from reportlab.lib.pagesizes import letter
-        from reportlab.pdfgen import canvas
-    except Exception as e:
-        raise ImportError(f"reportlab is required for PDF assembly: {e}")
-    images = [
-        f
-        for f in sorted(os.listdir(image_dir))
-        if f.lower().endswith((".png", ".jpg", ".jpeg", ".tif", ".tiff"))
-    ]
-    if not images:
-        raise RuntimeError(f"PDF: no images found in {image_dir}")
-    # Determine the output paths for the final PDF.  We produce two
-    # copies: one stored under ``output/music/<title>/<artist>/<key>/``
-    # organised by song, artist and key, and another under the log
-    # directory ``output/logs/<run_ts>/<title>/4_final_pdf`` for
-    # debugging.  Use metadata recorded during scraping to build a
-    # meaningful directory structure.  When metadata is missing, fall
-    # back to the provided ``output_pdf`` parameter.
-    final_pdf_path: str
-    debug_pdf_path: str
-    try:
-        meta = SCRAPE_METADATA.copy()
-        # Retrieve run timestamp for logs
-        run_ts = meta.get('run_ts') or os.environ.get('RUN_TS') or ''
-        # Use sanitised components for file and directory names.  The
-        # title stored in metadata is already sanitised.  If any
-        # component is missing, substitute a sensible default.
-        # Use unified sanitiser for title, artist, key and instrument names
-        title_meta = meta.get('title', '') or 'unknown'
-        artist_meta = meta.get('artist', '') or 'unknown'
-        key_meta = meta.get('key', '') or 'unknown'
-        instrument_meta = meta.get('instrument', '') or 'unknown'
-        title_dir = sanitize_title(title_meta)
-        artist_dir = sanitize_title(artist_meta)
-        key_dir = sanitize_title(key_meta)
-        instrument_part = sanitize_title(instrument_meta)
-        # Build final directory under output/music
-        pdf_root = os.path.join(os.getcwd(), "output", "music")
-        final_dir = os.path.join(pdf_root, title_dir, artist_dir, key_dir)
-        os.makedirs(final_dir, exist_ok=True)
-        # Compose file name
-        file_title = sanitize_title(title_meta.lower()) if title_meta else 'output'
-        file_name = f"{file_title}_{instrument_part}_{key_dir}.pdf"
-        final_pdf_path = os.path.join(final_dir, file_name)
-        # Build debug directory under logs mirroring the song/artist/key/instrument
-        # structure.  If a run timestamp is available, construct the full
-        # hierarchy; otherwise fall back to a directory adjacent to the
-        # input images.  The instrument component comes from instrument_part.
-        if run_ts:
-            debug_dir = os.path.join(
-                os.getcwd(),
-                "output",
-                "logs",
-                run_ts,
-                title_dir,
-                artist_dir,
-                key_dir,
-                instrument_part,
-                "4_final_pdf",
-            )
-        else:
-            # Fallback: use the parent of the input image directory
-            debug_dir = os.path.join(os.path.dirname(image_dir.rstrip(os.sep)), "4_final_pdf")
-        os.makedirs(debug_dir, exist_ok=True)
-        debug_pdf_path = os.path.join(debug_dir, file_name)
-    except Exception:
-        # Fallback: use provided output_pdf and create directories if necessary
-        final_pdf_path = output_pdf
-        debug_pdf_path = output_pdf
-        os.makedirs(os.path.dirname(final_pdf_path) or '.', exist_ok=True)
-    # Prepare the canvas for the PDF.  Render pages to the final
-    # destination first, then copy the resulting file into the debug
-    # directory.  This avoids rendering twice.
-    c = canvas.Canvas(final_pdf_path, pagesize=letter)
-    width, height = letter
-    for fname in images:
-        path = os.path.join(image_dir, fname)
-        c.drawImage(path, 0, 0, width=width, height=height, preserveAspectRatio=True)
-        c.showPage()
-    c.save()
-    # Copy the final PDF into the debug directory for debugging purposes
-    try:
-        if final_pdf_path != debug_pdf_path:
-            shutil.copyfile(final_pdf_path, debug_pdf_path)
-    except Exception as cp_err:
-        logger.error("ASSEMBLER: failed to copy PDF to debug directory: %s", cp_err)
-    logger.info("ASSEMBLER: wrote %d pages to %s", len(images), final_pdf_path)
-    logger.info("ASSEMBLER completed in %.3fs", time.perf_counter() - start)
-    # Do not clean up intermediate directories.  All stages are
-    # preserved under the run directory for future debugging and
-    # reproducibility.  Return the path to the assembled PDF.
-    return final_pdf_path
